% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox.R
\name{boxcox}
\alias{boxcox}
\title{Apply a box cox transformation}
\usage{
boxcox(data)
}
\arguments{
\item{x}{The vector to be transformed}

\item{plot_it}{Display a plot of \code{x}}
}
\value{
A vector that is transformed
}
\description{
This function helps make non-normal data more normally distributed via a box cox transformation
}
\examples{
boxcox(gapminder$lifeExp)

}
