% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_trans.R
\name{log_trans}
\alias{log_trans}
\title{Logarithmic transformation}
\usage{
log_trans(data, base = exp(1))
}
\arguments{
\item{base}{the base to perform log transformation}

\item{x}{The vector to be transformed}
}
\value{
A vector that is the:
\itemize{
     \item log (for \code{log})
}
 of \code{x}.
}
\description{
This function take the logarithmic of a vector
}
\examples{
log(9,3)
log(100,10)
log(27,3)

}
